apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5
openai==1.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
pypdf==4.2.0
pandas==2.2.2
numpy==1.26.4
tiktoken==0.7.0
tenacity==8.3.0
structlog==24.1.0
slowapi==0.1.9
limits==3.11.0
redis==5.0.4
email-validator==2.1.1
python-magic==0.4.27

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0

# Linting
ruff==0.4.3
